move hqq quantization to quant_primitives
